Rename package misc to utils
With our Pair implementation gone, we only have utility classes left in
this package.
