+++ src/main/java/de/uni_kl/cs/discodnc/minplus/dnc/affine/Deconvolution_DNC_Affine.java
@@ -36,8 +36,8 @@ import de.uni_kl.cs.discodnc.Calculator;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
-import de.uni_kl.cs.discodnc.misc.CheckUtils;
 import de.uni_kl.cs.discodnc.numbers.Num;
+import de.uni_kl.cs.discodnc.utils.CheckUtils;
 
 public abstract class Deconvolution_DNC_Affine {
