+++ src/main/java/de/uni_kl/cs/discodnc/nc/arrivalbounds/PmooArrivalBound.java
@@ -32,7 +32,6 @@ package de.uni_kl.cs.discodnc.nc.arrivalbounds;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
-import de.uni_kl.cs.discodnc.misc.SetUtils;
 import de.uni_kl.cs.discodnc.nc.AbstractArrivalBound;
 import de.uni_kl.cs.discodnc.nc.AnalysisConfig;
 import de.uni_kl.cs.discodnc.nc.AnalysisConfig.Multiplexing;
@@ -46,6 +45,7 @@ import de.uni_kl.cs.discodnc.network.Link;
 import de.uni_kl.cs.discodnc.network.Network;
 import de.uni_kl.cs.discodnc.network.Path;
 import de.uni_kl.cs.discodnc.network.Server;
+import de.uni_kl.cs.discodnc.utils.SetUtils;
 
 import java.util.Collections;
 import java.util.HashSet;
