+++ src/main/java/de/uni_kl/cs/discodnc/nc/arrivalbounds/PmooArrivalBound_SinkTreeTbRl.java
@@ -32,11 +32,11 @@ import de.uni_kl.cs.discodnc.Calculator;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
-import de.uni_kl.cs.discodnc.misc.SetUtils;
 import de.uni_kl.cs.discodnc.network.Flow;
 import de.uni_kl.cs.discodnc.network.Link;
 import de.uni_kl.cs.discodnc.network.Network;
 import de.uni_kl.cs.discodnc.network.Server;
+import de.uni_kl.cs.discodnc.utils.SetUtils;
 
 import java.util.HashMap;
 import java.util.HashSet;
