+++ src/main/java/de/uni_kl/cs/discodnc/utils/CheckUtils.java
@@ -1,4 +1,4 @@
-package de.uni_kl.cs.discodnc.misc;
+package de.uni_kl.cs.discodnc.utils;
 
 import java.util.Set;
